(1) refactored generic function.activation functions into neat, since they cannot be reused for gradient descent API's, (2) added random distribution API's, (3) fixed alpha zero inference algorithm, (4) fixed tic tac toe test case, (5) among other minor changes

+++ src/com/dipasquale/ai/rl/neat/core/ContextObjectSpeciationParameters.java
@@ -27,8 +27,8 @@ final class ContextObjectSpeciationParameters implements Context.SpeciationParam
     }
 
     @Override
-    public double compatibilityThreshold(final int generation) {
-        return compatibilityThreshold * Math.pow(compatibilityThresholdModifier, generation);
+    public float compatibilityThreshold(final int generation) {
+        return compatibilityThreshold * (float) Math.pow(compatibilityThresholdModifier, generation);
     }
 
     @Override
